feat: move index.jsx to tsx
