+++ src/index.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import ReactDOM, { Root } from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
+import { configureStore } from '@reduxjs/toolkit';
+import { Provider } from 'react-redux';
+
+import './index.css';
+import { rootReducer } from 'services/reducers/index.js';
+
+import App from 'components/app/app.jsx';
+
+const store = configureStore({
+  reducer: rootReducer,
+  devTools: process.env.NODE_ENV !== 'production'
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
+
+const rootElement = document.getElementById('root') as HTMLElement | null;
+
+if (rootElement) {
+  const root: Root = ReactDOM.createRoot(rootElement);
+
+  root.render(
+    <React.StrictMode>
+      <Provider store={store}>
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
+      </Provider>
+    </React.StrictMode>
+  );
+} else {
+  console.error("Element with id 'root' not found in the DOM.");
+}
